+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### vNext
 
+- [DAV-3734] Add in-wallet transaction modals
 - [DAV-3669] Add asset transfer (ethereum)
 - [DAV-3767] Add GTM integration and click IDs